Fixing imports and renaming files
