+++ questions/index.js
@@ -2,7 +2,7 @@
 const express = require('express');
 const cors = require('cors');
 // Routes:
-const questionRoutes = require('./routes/question-routes.js');
+const questionRoutes = require('./routes/question-routes-api.js');
 
 // App definition and
 const app = express();
