+++ users/routes/auth-routes.js
@@ -1,7 +1,7 @@
 const express = require('express');
 const router = express.Router();
 const bcrypt = require('bcrypt');
-const { User } = require('../models/user-model');
+const { User } = require('../services/user-model');
 
 
 require('dotenv').config();
