+++ users/routes/user-routes-api.js
@@ -1,6 +1,6 @@
 const express = require('express');
 const router = express.Router();
-const { User } = require('../models/user-model');
+const { User } = require('../services/user-model');
 
 //Get all users
 router.get('/allUsers', async (req,res) => {
